refactor: rename product to product_model
To clearly separate it from Product.jsx file
